% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bbar.R
\name{Bbar}
\alias{Bbar}
\title{Given a tree, return the average root to tip distance between a random pair of taxa. For calculating ICC the phylogenetic variance
should always be multiplied by the average root-to-tip branch length if you want the average ICC. If the tree is ultrametric and scaled then the
root-to-tip branch length is one and there is no need.}
\usage{
Bbar(phylo, type = c("All", "Genus", "Species"))
}
\arguments{
\item{phylo}{object of class phylo}

\item{type}{at what level calculation should be at, between all species, at the level of genus or any combination within genus}
}
\description{
Given a tree, return the average root to tip distance between a random pair of taxa. For calculating ICC the phylogenetic variance
should always be multiplied by the average root-to-tip branch length if you want the average ICC. If the tree is ultrametric and scaled then the
root-to-tip branch length is one and there is no need.
}
\examples{
# requires ape
library(ape)
tree <- rtree(100, rooted = TRUE)
Bbar(tree, type = "Genus")
}
\keyword{covariance,}
\keyword{tree}
\keyword{variance,}
