% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pvals.R
\name{calc_pvals}
\alias{calc_pvals}
\title{Calculating p-values for the annoying lme4 Gaussian mixed model}
\usage{
calc_pvals(lmermod)
}
\arguments{
\item{lmermod}{model of class lmerMod}
}
\description{
Calculating p-values for the annoying lme4 Gaussian mixed model
}
\examples{
# needs data.table
library(data.table)
mod <- lmer(y ~ x + (1 | z))
calc_pvals(mod)
}
\keyword{effects,}
\keyword{fixed}
\keyword{lme4,}
\keyword{p-values}
