% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densMode.R
\name{densMode}
\alias{densMode}
\title{Calculate x and y axis for peak density on MCMC objects (http://ianmadd.github.io/pages/PeakDensityDistribution.html)}
\usage{
densMode(x)
}
\arguments{
\item{x}{a vector of values from which to calculate a density}
}
\description{
Calculate x and y axis for peak density on MCMC objects (http://ianmadd.github.io/pages/PeakDensityDistribution.html)
}
\examples{
x <- c(1,2,3,2,2,2,2,1,2,12,2,2)
densMode(x)
}
\keyword{ICC,}
\keyword{MCMCglmm,}
\keyword{explained}
\keyword{repeatability,}
\keyword{variance}
