% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lmerfixplot.R
\name{g_lmerfixplot}
\alias{g_lmerfixplot}
\title{ggplot2 ouptut of Tukey comparisons for lmer models}
\usage{
g_lmerfixplot(model, factor = list(...), reorder = FALSE)
}
\arguments{
\item{model}{model of class lmerMod}

\item{factor}{list of the form list(pairwise ~ your variable of interest)}

\item{reorder}{should the plot be reordered from high to low effect size?}
}
\description{
ggplot2 ouptut of Tukey comparisons for lmer models
}
\examples{
mod <- lmer(y ~ x + (1 | z))
g_lmerfixplot(mod, list(pairwise ~ x), reorder = TRUE)
}
\keyword{effects,}
\keyword{fixed}
\keyword{lme4,}
\keyword{plot,}
\keyword{tukey}
