% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCfixplot.R
\name{MCMCfixplot}
\alias{MCMCfixplot}
\title{ggplot2 output of the fixed effects from a MCMCglmm model}
\usage{
MCMCfixplot(mod)
}
\arguments{
\item{mod}{model of class MCMCglmm}
}
\description{
ggplot2 output of the fixed effects from a MCMCglmm model
}
\examples{
# requires ggplot2
library(ggplot2)
mod <- MCMCglmm(y ~ x1 + x2, random =  ~ z)
MCMCfixplot(mod)
}
\keyword{MCMCglmm,}
\keyword{effects,}
\keyword{fixed}
\keyword{plot}
