% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCranef.R
\name{MCMCranef}
\alias{MCMCranef}
\title{Visualise the random effect posterior distributions}
\usage{
MCMCranef(mod, group = FALSE, data = FALSE)
}
\arguments{
\item{mod}{model of class MCMCglmm}

\item{group}{logical, should the random effects be grouped?}
}
\description{
Visualise the random effect posterior distributions
}
\examples{
# requires plyr, dplyr,reshape2, ggplot2
library(plyr); library(tidyverse)
model <- MCMCglmm(y ~ x, random = ~ z)
MCMCranef(model)
}
\keyword{MCMCglmm,}
\keyword{modes}
\keyword{posterior}
