% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCRepnorm2.R
\name{MCMCRepnorm2}
\alias{MCMCRepnorm2}
\title{Calculate variance explained by a random variable (repeatability or Intra-Class Coefficient) on the data scale for Gaussian models}
\usage{
MCMCRepnorm2(mod)
}
\arguments{
\item{mod}{model of class MCMCglmm}
}
\description{
Calculate variance explained by a random variable (repeatability or Intra-Class Coefficient) on the data scale for Gaussian models
}
\examples{
model <- MCMCglmm(y ~ x, random = ~ z)
MCMCRepnorm2(model, "z")
}
\keyword{ICC,}
\keyword{MCMCglmm,}
\keyword{explained}
\keyword{repeatability,}
\keyword{variance}
