% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCReppois.R
\name{MCMCReppois}
\alias{MCMCReppois}
\title{Calculate variance explained by a random variable (repeatability or Intra-Class Coefficient) on the link scale}
\usage{
MCMCReppois(mod, y = "variable")
}
\arguments{
\item{mod}{model of class MCMCglmm}

\item{y}{character vector of length 1, identical to name of random effect.}
}
\description{
Calculate variance explained by a random variable (repeatability or Intra-Class Coefficient) on the link scale
}
\examples{
model <- MCMCglmm(y ~ x, random = ~ z)
MCMCReppois(model, "z")
}
\keyword{ICC,}
\keyword{MCMCglmm,}
\keyword{explained}
\keyword{repeatability,}
\keyword{variance}
