% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.sdev.R
\name{posterior.sd}
\alias{posterior.sd}
\title{Posterior standard deviations for fixed (and random) effects}
\usage{
posterior.sd(model, pr = FALSE)
}
\arguments{
\item{model}{model of class MCMCglmm}

\item{pr}{logical. Should the posterior standard deviations of the random effects be calculated? Option pr = TRUE must be in the MCMCglmm model.}
}
\description{
Posterior standard deviations for fixed (and random) effects
}
\keyword{MCMCglmm,}
\keyword{deviation}
\keyword{effects,}
\keyword{fixed}
\keyword{random}
\keyword{standard}
