% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solapply.R
\name{Solapply}
\alias{Solapply}
\title{Apply a function over each of the stored posterior solutions from a MCMCglmm Model}
\usage{
Solapply(model, FUN = posterior_mode, ...)
}
\arguments{
\item{model}{model of class MCMCglmm}

\item{FUN}{function name to apply over object, default is posterior mode, suggested also is coda::HPDinterval}
}
\description{
Apply a function over each of the stored posterior solutions from a MCMCglmm Model
}
\examples{
# requires data.table
library(data.table)
model <- MCMCglmm(y ~ x, random = ~ z)
Solapply(model, coda::HPDinterval)
}
\keyword{MCMCglmm,}
\keyword{modes}
\keyword{posterior}
