% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_decimal.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Specify the number of decimal places
Please see (https://stackoverflow.com/questions/3443687/formatting-decimal-places-in-r)
for the function conceptions}
\usage{
specify_decimal(x, k)
}
\arguments{
\item{x}{decimal to convert}

\item{k}{number of decimal places}
}
\description{
Specify the number of decimal places
Please see (https://stackoverflow.com/questions/3443687/formatting-decimal-places-in-r)
for the function conceptions
}
\examples{
x <- 0.324153245
specify_decimal(x, 3)
}
\keyword{decimal}
\keyword{places,}
