% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{traces}
\alias{traces}
\title{Lattice output of the MCMC chains from the fixed and random effects from a MCMCglmm model}
\usage{
traces(mod)
}
\arguments{
\item{mod}{model of class MCMCglmm}
}
\description{
Lattice output of the MCMC chains from the fixed and random effects from a MCMCglmm model
}
\examples{
mod <- MCMCglmm(y ~ x1 + x2, random =  ~ z)
traces(mod)[[1]]; traces(mod)[[2]]
}
\keyword{MCMCglmm,}
\keyword{effects,}
\keyword{fixed}
\keyword{plot}
