% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultm.R
\name{ultm}
\alias{ultm}
\title{Convert a phylo object to an inverse distance matrix that MCMCglmm understands}
\usage{
ultm(tree, MPL = FALSE, inverseA = TRUE)
}
\arguments{
\item{tree}{tree of class phylo}

\item{MPL}{logical, should the function compute the mean path length?}

\item{inverseA}{logical, should the inverse matrix be computed?}
}
\description{
Convert a phylo object to an inverse distance matrix that MCMCglmm understands
}
\examples{
tree <- ape::read.tree("./tree.newick") # newick format
ultm(tree, MPL = FALSE, inverseA = TRUE)
}
\keyword{MCMCglmm,}
\keyword{ape,}
\keyword{phylo,}
\keyword{tree}
