% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarExpl.mer.R
\name{VarExpl.mer}
\alias{VarExpl.mer}
\title{Calculate variance explained from a lme4 model}
\usage{
VarExpl.mer(mod)
}
\arguments{
\item{mod}{model of class (g)lmerMod}
}
\description{
Calculate variance explained from a lme4 model
}
\examples{
mod <- lmer(y ~ x + (1 | z))
VarExpl.mer(mod)
}
\keyword{explained,}
\keyword{lme4,}
\keyword{repeatability}
\keyword{variance}
