% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCVdensity.R
\name{VCVdensity}
\alias{VCVdensity}
\title{Visualise the VCV of a MCMCglmm model}
\usage{
VCVdensity(mod)
}
\arguments{
\item{mod}{model of class MCMCglmm}
}
\description{
Visualise the VCV of a MCMCglmm model
}
\examples{
# requires ggplot2
library(ggplot2)
model <- MCMCglmm(y ~ x, random = ~ z)
VCVdensity(model)
}
\keyword{MCMCglmm,}
\keyword{modes}
\keyword{posterior}
