% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald.test.R
\name{Wald.test}
\alias{Wald.test}
\title{Performing Wald tests on MCMCglmm fixed effects}
\usage{
Wald.test(mod, fixed)
}
\arguments{
\item{mod}{model of class MCMCglmm}

\item{fixed}{numeric vector specifying terms to be tested. In summary(mod), in location effects, the intercept is 1. Jointly test a factor by including all levels (minus intercept).}
}
\description{
Performing Wald tests on MCMCglmm fixed effects
}
\examples{
# needs aod
library(aod)
mod <-  MCMCglmm(y ~ x, random = ~ z)
# if x is a 4 levelled factor
Wald.test(mod = mod, fixed = 2:4)
}
\keyword{MCMCglmm,}
\keyword{Wald}
\keyword{effects,}
\keyword{fixed}
\keyword{p-values,}
\keyword{test}
