% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bbar.R
\name{Bbar}
\alias{Bbar}
\title{Given a tree, return the average root to tip distance between a random pair of taxa}
\usage{
Bbar(phylo, type = c("All", "Genus", "Species"))
}
\arguments{
\item{phylo}{object of class phylo}

\item{type}{at what level calculation should be at, between all species, at the level of genus or any combination within genus}
}
\description{
Given a tree, return the average root to tip distance between a random pair of taxa
}
\examples{
# requires ape
library(ape)
tree <- rtree(100, rooted = TRUE)
Bbar(tree, type = "Genus")
}
\keyword{covariance,}
\keyword{tree}
\keyword{variance,}
